backend/django_corpciti/serviciosCorpciti/serializers.py
from .models import Cliente, Pago, Asesor, Congreso_seminario, Factura
from rest_framework import serializers

class ClienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cliente
        fields = '__all__'

class PagoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Pago
        fields = '__all__'

class AsesorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Asesor
        fields = '__all__'

class CongresoSerializer(serializers.ModelSerializer):
    class Meta:
        model = Congreso_seminario
        fields = '__all__'

class FacturaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Factura
        fields = '__all__'